// src/core/types/User.ts
export type UserResponse = {
    id: number;
    login: string;
    name: string;
    company: string;
    blog: string;
    location: string;
    twitter_username: string;
    public_repos: number;
    followers: number;
    following: number;
    created_at: string;
    avatar_url: string;
    html_home: string;
}